+++ AgroAgil/app/src/main/java/com/example/agroagil/Cultivo/ui/CultivoTypeInfoEditScreen.kt
@@ -327,7 +327,7 @@ fun CultivoTypeInfoEdit(cultivoViewModel: CultivoViewModel, stockViewModel: Stoc
                                     if (price) {
                                         if (selectedOptionText == "Valor") {
                                             selectedCrop.forEach {
-                                                it.price = priceText.toFloat().toDouble()
+                                                it.price = priceText.toFloat()
                                             }
                                         } else {
                                             selectedCrop.forEach {
